Add wrapper method for apply resource
