+++ internal/kubernetes/interface.go
@@ -1,0 +1,5 @@
+package kubernetes
+
+type Client interface {
+	ApplyResource([]byte) error
+}
